# Precompute w to avoid recomputing it over and over